feat: abstracting code to smaller files
This will help start testing all the stuff
