+++ package.json
@@ -7,12 +7,15 @@
   "dependencies": {
     "@types/react": "16.7.18",
     "@types/react-dom": "16.0.11",
-    "morsee": "1.0.1",
+    "codacy-coverage": "3.3.0",
+    "jest": "23.6.0",
+    "morsee": "1.0.2",
     "react": "next",
     "react-dom": "next",
     "react-scripts": "2.0.3"
   },
   "devDependencies": {
+    "@types/jest": "23.3.10",
     "@types/react": "16.7.18",
     "@types/react-dom": "16.0.11"
   },
