+++ src/index.spec.tsx
@@ -1,0 +1,5 @@
+describe('giraffe', () => {
+  test('says', () => {
+    expect('hello world').toBe('hello world');
+  });
+});
