+++ src/utils/enum.ts
@@ -1,0 +1,3 @@
+const morseCodeStrings: string[] = ['.', '-'];
+
+export { morseCodeStrings };
