+++ src/utils/helper.ts
@@ -1,0 +1,8 @@
+import { morseCodeStrings as arr } from './enum';
+
+const compareInput = (valueTyped: string) =>
+  arr.some((v: string) => {
+    return !!valueTyped.includes(v);
+  });
+
+export { compareInput };
